+++ packages/auth/src/next/common/types.ts
@@ -1,0 +1,62 @@
+import type { GenericAuthWallet } from "@thirdweb-dev/wallets";
+import { z } from "zod";
+
+import { Json, LoginPayloadOutputSchema, User } from "../../core";
+
+export const PayloadBodySchema = z.object({
+  address: z.string(),
+  chainId: z.string().optional(),
+});
+
+export const ActiveBodySchema = z.object({
+  address: z.string(),
+});
+
+export const LoginPayloadBodySchema = z.object({
+  payload: LoginPayloadOutputSchema,
+});
+
+export type ThirdwebAuthRoute =
+  | "payload"
+  | "login"
+  | "logout"
+  | "user"
+  | "switch-account";
+
+export type ThirdwebAuthUser<
+  TData extends Json = Json,
+  TSession extends Json = Json,
+> = User<TSession> & {
+  data?: TData;
+};
+
+export type ThirdwebAuthConfigShared = {
+  domain: string;
+  wallet: GenericAuthWallet;
+  authOptions?: {
+    statement?: string;
+    uri?: string;
+    version?: string;
+    chainId?: string;
+    resources?: string[];
+    validateNonce?:
+      | ((nonce: string) => void)
+      | ((nonce: string) => Promise<void>);
+    validateTokenId?:
+      | ((tokenId: string) => void)
+      | ((tokenId: string) => Promise<void>);
+    loginPayloadDurationInSeconds?: number;
+    tokenDurationInSeconds?: number;
+    refreshIntervalInSeconds?: number;
+  };
+  cookieOptions?: {
+    domain?: string;
+    path?: string;
+    sameSite?: "lax" | "strict" | "none";
+    secure?: boolean;
+  };
+};
+
+export type ThirdwebNextContext = {
+  params?: Record<string, string | string[]>
+};